+++ test/ecto/range_operators_test.exs
@@ -1,0 +1,50 @@
+defmodule Ecto.RangeOperatorsTest do
+  # @related [subject](/lib/ecto/range_operators.ex)
+  @moduledoc false
+  use Test.DataCase, async: true
+  import Ecto.Query
+  import Ecto.RangeOperators
+  import Ecto.UTCDateTimeRange, only: [sigil_t: 2]
+  alias Ecto.UTCDateTimeRange
+
+  doctest Ecto.RangeOperators
+
+  deftemptable :things_with_time_ranges do
+    column(:during, :tstzrange)
+    column(:tid, :string)
+  end
+
+  setup do
+    create_temp_tables()
+    :ok
+  end
+
+  defmodule Thing do
+    @moduledoc false
+    use Ecto.Schema
+
+    schema "things_with_time_ranges_temp" do
+      field(:during, UTCDateTimeRange)
+      field(:tid, :string)
+    end
+
+    def changeset(struct \\ %__MODULE__{}, attrs),
+      do: struct |> cast(attrs, [:during, :tid])
+  end
+
+  describe "contains" do
+    test "queries for records where the given timestamp is contained in a record's range" do
+      Thing.changeset(%{during: ~t{2020-02-02T13:00:00Z - 2020-02-02T14:00:00Z}, tid: "early"}) |> Test.Repo.insert()
+      Thing.changeset(%{during: ~t{2020-02-02T14:30:00Z - 2020-02-02T15:00:00Z}, tid: "middle"}) |> Test.Repo.insert()
+      Thing.changeset(%{during: ~t{2020-02-02T15:00:00Z - 2020-02-02T15:30:00Z}, tid: "later"}) |> Test.Repo.insert()
+
+      time = ~U[2020-02-02T14:45:00Z]
+
+      from(_ in Thing, as: :things)
+      |> where([things: things], contains(things.during, ^time))
+      |> Test.Repo.all()
+      |> tids()
+      |> assert_eq(["middle"])
+    end
+  end
+end